services/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Daniel Nurieli <daniel_nurieli@alumni.brown.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pydantic-settings = "^2.10.1"
firebase-admin = "^7.1.0"
sqlalchemy = "^2.0.43"
asyncpg = "^0.30.0"
alembic = "^1.16.5"
python-dotenv = "^1.1.1"
httpx = "^0.28.1"
tenacity = "^9.1.2"
google-cloud-secret-manager = "^2.20.0"
PyJWT = "^2.10.0"
python-multipart = "^0.0.20"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
